Fixed to execute multi test cases
